use dtype= in call to empty in ghost

+++ dask/array/ghost.py
@@ -362,7 +362,7 @@ def add_dummy_padding(x, depth, boundary):
     >>> import dask.array as da
     >>> x = da.arange(6, chunks=3)
     >>> add_dummy_padding(x, {0: 1}, {0: 'none'}).compute()  # doctest: +NORMALIZE_WHITESPACE
-    array([..., 0., 1., 2., 3., 4., 5., ...])
+    array([..., 0, 1, 2, 3, 4, 5, ...])
     """
     for k, v in boundary.items():
         if v == 'none':
@@ -374,7 +374,7 @@ def add_dummy_padding(x, depth, boundary):
             empty_chunks = list(x.chunks)
             empty_chunks[k] = (d,)
 
-            empty = wrap.empty(empty_shape, chunks=empty_chunks)
+            empty = wrap.empty(empty_shape, chunks=empty_chunks, dtype=x.dtype)
 
             out_chunks = list(x.chunks)
             ax_chunks = list(out_chunks[k])
